+++ code/game/objects/items/weapons/table_rack_parts.dm
@@ -15,11 +15,12 @@
  * Table Parts
  */
 // Return TRUE if reacted to a tool.
+/obj/item/weapon/table_parts
+	var/build_time = 0
+
 /obj/item/weapon/table_parts/proc/attack_tools(obj/item/W, mob/user)
-	if(iswrench(W))
-		new /obj/item/stack/sheet/metal( user.loc )
-		//SN src = null
-		qdel(src)
+	if(iswrenching(W))
+		deconstruct(TRUE, user)
 		return TRUE
 
 	else if(istype(W, /obj/item/stack/rods))
@@ -39,25 +40,41 @@
 
 	return ..()
 
+/obj/item/weapon/table_parts/deconstruct(disassembled, user = FALSE)
+	if(flags & NODECONSTRUCT)
+		return ..()
+	var/turf/T = get_turf(user || src)
+	for(var/debrit_type in debris)
+		new debrit_type(T)
+	..()
+
 /obj/item/weapon/table_parts/attack_self(mob/user)
 	var/turf/simulated/T = get_turf(user)
-	if (T.CanPass(null, T))
-		var/obj/structure/table/R = new table_type( T )
-		to_chat(user, "<span class='notice'>You assemble [src].</span>")
-		R.add_fingerprint(user)
-		qdel(src)
-	else
+	if(!can_place(T))
 		to_chat(user, "<span class='warning'>You can't put it here!</span>")
+		return
+	if(build_time > 0 && !handle_fumbling(user, src, build_time, list(/datum/skill/engineering = SKILL_LEVEL_NOVICE)))
+		return
+	if(!can_place(T))
+		to_chat(user, "<span class='warning'>You can't put it here!</span>")
+		return
+	var/obj/structure/table/R = new table_type(T)
+	to_chat(user, "<span class='notice'>You assemble [src].</span>")
+	R.add_fingerprint(user)
+	qdel(src)
 
+/obj/item/weapon/table_parts/proc/can_place(turf/T)
+	return T && T.CanPass(null, T)
 
 /*
  * Reinforced Table Parts
  */
+/obj/item/weapon/table_parts/reinforced
+	build_time = SKILL_TASK_AVERAGE
+
 /obj/item/weapon/table_parts/reinforced/attack_tools(obj/item/W, mob/user)
-	if(iswrench(W))
-		new /obj/item/stack/sheet/metal(user.loc)
-		new /obj/item/stack/rods(user.loc)
-		qdel(src)
+	if(iswrenching(W))
+		deconstruct(TRUE, user)
 		return TRUE
 	return FALSE
 
@@ -65,9 +82,8 @@
  * Glass Table Parts
  */
 /obj/item/weapon/table_parts/glass/attack_tools(obj/item/W, mob/user)
-	if(iswrench(W))
-		new /obj/item/stack/sheet/glass( user.loc )
-		qdel(src)
+	if(iswrenching(W))
+		deconstruct(TRUE, user)
 		return TRUE
 	return FALSE
 
@@ -76,9 +92,8 @@
  * Wooden Table Parts
  */
 /obj/item/weapon/table_parts/wood/attack_tools(obj/item/W, mob/user)
-	if(iswrench(W))
-		new /obj/item/stack/sheet/wood(user.loc)
-		qdel(src)
+	if(iswrenching(W))
+		deconstruct(TRUE, user)
 		return TRUE
 
 	else if(istype(W, /obj/item/stack/tile/grass))
@@ -95,9 +110,8 @@
  * Fancy Wooden Table Parts
  */
 /obj/item/weapon/table_parts/wood/fancy/attack_tools(obj/item/W, mob/user)
-	if(iswrench(W))
-		new /obj/item/stack/sheet/wood(user.loc)
-		qdel(src)
+	if(iswrenching(W))
+		deconstruct(TRUE, user)
 		return TRUE
 	return FALSE
 
@@ -106,10 +120,8 @@
  */
 
 /obj/item/weapon/table_parts/wood/poker/attack_tools(obj/item/W, mob/user)
-	if(iswrench(W))
-		new /obj/item/stack/sheet/wood(user.loc)
-		new /obj/item/stack/tile/grass(user.loc)
-		qdel(src)
+	if(iswrenching(W))
+		deconstruct(TRUE, user)
 		return TRUE
 	return FALSE
 
@@ -117,11 +129,15 @@
  * Rack Parts
  */
 /obj/item/weapon/rack_parts/attackby(obj/item/I, mob/user, params)
-	if(iswrench(I))
-		new /obj/item/stack/sheet/metal( user.loc )
-		qdel(src)
+	if(iswrenching(I))
+		deconstruct(TRUE, user)
 		return
 	return ..()
+
+/obj/item/weapon/rack_parts/deconstruct(disassembled, user = FALSE)
+	if(!(flags & NODECONSTRUCT))
+		new /obj/item/stack/sheet/metal(get_turf(user || src))
+	..()
 
 /obj/item/weapon/rack_parts/attack_self(mob/user)
 	var/turf/simulated/T = get_turf(user)
